requirements.txt
requests
python-dotenv
fastapi
uvicorn[standard]

